+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+1.1.0.0
+-----
+* Odtwarzanie dźwięku pierwszej nowej wiadomości.
+* Blokowanie odtwarzania domyślnych dźwięków nowej wiadomości zamiast ich wyłączania.
+* Poprawne pobieranie uchwytu do okna rozmowy przy jego otwieraniu.
+
 1.0.2.0
 -----
 * Poprawne wykrywanie aktywnego okna.